Move redact to archiver, and refactor checksums editing
