+++ src/rachis/sdk/result.py
@@ -798,72 +798,6 @@ class Artifact(Result):
     def validate_checksums(self):
         super().validate()
 
-    def get_provenance_path(self):
-        return os.path.join(self._archiver.root_dir, 'provenance')
-
-    def redact_metadata(self):
-        '''
-        Finds the filepath to every metadata file based on the name included
-        in the `action.yaml` files, then removes these paths. It is also
-        neccessary to remove the corresponding lines in the `checksums` and
-        `action.yaml` files.
-        '''
-
-        def get_metadata_objects(yaml_dict):
-            '''
-            Recursively searches through dictionary returned by
-            `yaml.safe_load` returning any `MetadataInfo` objects.
-            '''
-            metadata = []
-            if isinstance(yaml_dict, MetadataInfo):
-                metadata.append(yaml_dict)
-            elif isinstance(yaml_dict, dict):
-                for value in yaml_dict.values():
-                    metadata.extend(get_metadata_objects(value))
-            elif isinstance(yaml_dict, (list, tuple)):
-                for value in yaml_dict:
-                    metadata.extend(get_metadata_objects(value))
-            return metadata
-
-        yaml.SafeLoader.add_constructor('!metadata', metadata_path_constructor)
-        prov_path = self.get_provenance_path()
-
-        metadata_paths = []
-        relative_metadata_paths = []
-        for action_yaml in Path(prov_path).rglob('action.yaml'):
-            with open(action_yaml) as fh:
-                metadata = yaml.safe_load(fh)
-                metadatas = get_metadata_objects(metadata)
-                if metadatas is None:
-                    raise ValueError(
-                        'Cannot redact metadata from an artifact that does '
-                        'contain metadata.'
-                    )
-                for metadata in metadatas:
-                    metadata_paths.append(
-                        action_yaml.parent / metadata.relative_fp
-                    )
-                    relative_metadata_paths.append(metadata.relative_fp)
-
-        for metadata_path in metadata_paths:
-            os.remove(metadata_path)
-
-        # Filter out lines refering to the now missing metadata files
-        for file in Path(prov_path).parent.iterdir():
-            if 'checksums' in file.name:
-                with open(file, 'r') as f:
-                    lines = f.readlines()
-
-                lines = [
-                    line for line in lines if not any(
-                        r_path in line for r_path in relative_metadata_paths
-                    )
-                ]
-
-                with open(file, 'w') as f:
-                    f.writelines(lines)
-
-
 
 class Visualization(Result):
     extension = '.qzv'
